hotfix: fix minion and chuffed always rebuilding
We were using the .parse_callbacks(Box::new(bindgen::CargoCallbacks))
option of bindgen to invalidate the build based on included files.

We always re-run ./configure in build.sh - this was invalidating the
build because the configure script updates the last modified timestamps
of generated headerfiles, even if the file contents themselves have not
changed.

We already rebuild on C++ changes as both chuffed and minion have
rerun-if-changed=vendor, rerun-if-changed=wrapper.h added manually.
Therefore, the bindgen functionality that does this is not needed. We
can assume if vendor/ has changed, then the configure generated headers
have changed - we do not need to track these directly.