make rule macro safe to use everywhere, and move rules into conjure_oxide.
This commit makes the rule macro safe to use in all crates, not just
conjure_rules. This allows for completely decentralised definition of
rules, without the user having to worrying about dependencies.

Previously, the macro expansion made imports from external crates, which
would break when used in crates that did not explicitly list said crates
in their Cargo.toml. Some simple solutions to this are to limit use of
the macro to a single crate, or to force users to add stuff to their
Cargo.toml.

These are both non-ideal: the former limits how "decentralised" rule
definition could be; the latter is a leaky abstraction and would make
updating this crate difficult for a user.

See:

- https://users.rust-lang.org/t/proc-macros-using-third-party-crate/42465/4
- https://doc.rust-lang.org/reference/procedural-macros.html#procedural-macro-hygiene

Procedural macros and external crate imports
==============================================

Procedural macros are unhygienic - they directly substitute into source
code, and do not have their own scope, imports, and so on.

Therefore, we cannot assume the user has any dependencies apart from the
one they imported the macro from. (Also, note Rust does not bring
transitive dependencies into scope, so we cannot assume the presence of
a dependency of the crate.)

To solve this, the crate the macro is in must re-export everything the
macro needs to run.

However, proc-macro crates can only export proc-macros. Therefore, we
must use a "front end crate" to re-export both the macro and all the
things it may need.

Conjure_rules is the frontend crate, and re-exports the macro from
conjure_rules_proc_macro.