add stats and solver run stats to global context
Add Stats and SolverStats objects to store global statistics within the
Context.

SolverStats contains the statistics for a single solver run: both those
from the SolverAdaptor, and those generated generically within the
Solver state machine (eg wall time). The stats from the SolverAdaptor
are merged with Solver's walltime using SolverStats::with_timings(..).

Solver::save_stats_to_context() appends these to an array in the global
Stats object within the Context. This is an array because multiple runs
of a solver could be done within one execution context. Note that adding
these stats globally is opt-in not opt-out.

Taking inspiration from Savile Row, we realised that SolverStats could
just be a flat "map" of optional fields and not have a solver
independent arbitrary nested representation. (as discussed and planned
prevously).

While we were originally planning multiple solver specific stats
representations held together by a trait, it is easier to develop on and
use if we just have a big list of optional fields instead, even if not
all solvers use them. (Savile Row uses a HashMap here, but I would
prefer fields to be typed and not need to use string casting
everywhere). The trait approach would've also likely involved
downcasting or a big switch statement somewhere!

Prerequisite changes to the context also added in this commit:

* Context should be a pointer not a value.
* Context is created before parsing, not afterwards.
* Context needs to be passed from Model to Solver, so that Solver has a
  reference to it.

Co-authored-by: Georgii Skorokhod <skorokhod.g@gmail.com>