ast: use new Uniplate types for AST traversal
* Use a manual implementation of Uniplate on Expression instead of the
  derive macro.

* Use the new uniplate::biplate::Uniplate type instead of
  uniplate::uniplate::Uniplate. This will be the only Uniplate type in
  the future.

These changes are to help development of the new Uniplate derive macro.

In https://github.com/conjure-cp/conjure-oxide/pull/290, I am in the
process of removing the old uniplate type (uniplate::uniplate::Uniplate)
and rewriting the macro - the changes made in this commit ensure that
during development of this PR Conjure Oxide still compiles
and passes the tests.