use a specific version of uniplate, not the latest version
This change makes conjure_core and conjure_oxide use the version of
uniplate found at a specific commit of this repo, instead of the latest
"path" version.

Doing this allows me to add breaking changes to uniplate easier without
failing CI -- instead of having to fix conjure_oxide for every breaking
change in uniplate, I can leave this till the end and do it all in one
go.