tests: add .eprime file extension support to the test runner
Add support for Essence Prime files (.eprime) to the test runner. This
change will help test Conjure Oxide as a replacement for Savile Row (the
current Essence Prime toolchain).

As Essence Prime is a subset of Essence, no extra work will be needed in
Conjure Oxide to support Essence Prime.