add partial evaluator
For expressions containing both constants and nested expressions, the partial
evaluator collapses constant values into a single value. For example, x + 1 + y
+ 2 + 3 ~> x + y + 6.

In some cases, the partial evaluator replaces the constraint with a truth
value. Currently, this is done when two identical constants are in an
all-different constraint, and for the boolean clauses ` M or true` and `M and
false`.

The difference between this and the existing evaluator is that the latter only
evaluates expressions made entirely of constants. Having both is useful: the
total evaluator is useful for solutions checking (amongst other things), and
the partial evaluator for model Optimisation.