change rule priorities to run simplifiers first; add reify
Create a rule priority taxonomy, and refactor rules to follow this new
convention. In particular, ensure that partial and total evaluation and
simplification rules happen before representation changes and solver
specific rules.

DESCRIPTION

Simplifying a model before applying representation changes should ensure
that nested sums do not get propagated across the model. For example,
the transformation for min turns it into many binary constraints. If
simplification was not done first, simplification steps would have to be
ran on each of these binary constraints, instead of just on the input
model.

Ensuring that everything is always in canonical forms should stop rules
looping, and help with pattern matching. The implementation of
canonical forms is incomplete and outside the scope of this patch.

These rule changes caused basic-div-04 to fail. To fix this, this patch
also does the following:

- add reify constraint (reify(X,1) if constraint X is SAT, reify(X,0) if
  constraint X is UNSAT)

- add rule not(X) ~> reify(X,0) (where X is a constraint), allowing
  negated constraints to be passed to Minion.

MOTIVATION

The commit c2b74602c (add broken complex partial evaluator test
(arithmetic, nesting, min), 2024-10-12) added a test case for the
partial evaluator that never terminated.

Simplification steps should be ran before the decomposition of the min
constraint and the conversion of expressions into more specialised /
solver specific forms (such as ineq, sumgeq, etc); however, this was not
the case in this test. In this test, ineq / sumgeq were being applied
before the sums had been flattened.

After this refactor, this test now terminates with failure. A flattening
rule for ineq is needed before this test passes and will be added
in a later commit.