WIP: minion: add flattening rule for binary operators
Replace the existing div_flattening rule with a more general one, useful
for all binary operators.

The flattening rule turns:

c `op1` (a `op2` b) ~> c `op1` x, x =
  where a and b are constants or variables.

When a and b are not constants or variables, I expect flattening to
happen recursively bottom up.

This rule currently only applies to division, as this is our only binary
operator; however, it is general enough to be extended to minus,
product, power, etc. when implemented.

Nested divisions are now supported, as demonstrated in div_undefzero-03-nested.