tools: add watch-tests.sh to rerun integration tests on changes.
Add a script watch-tests.sh to rerun the integration tests on changes to
the code or essence/eprime files. This gives a summary of failing tests
that is updated only when the source-code is.

While similar to bacon, bacon ignores ACCEPT=true, while this script
recognises it. Also, the output of this script is more readable, as it
uses the output of test-summary.sh not cargo test.