rules: move negated eq/neq rules from Minion ruleset to Base
Move not(eq(...)) ~> neq and not(neq(...)) ~> eq rules into base
ruleset, and raise their priorities to that of a simplification.

Both of these simplifications are Essence, solver-independent ones, not
Minion specific.

As discussed with Oz, despite `not(eq)` being similar to `neq`, `neq` is
a useful expression to use at "Essence level" due to undefinedness
semantics.