Changed Model constraints and Reduction new_top to Vec<Expression>
Major benefit is being close to the input model, which in turn means if we pretty print the model it would look closer to what the user wrote.

Refactored apply() function to target a specific constraint by index (constraint_idx) instead of modifying the entire expression tree.

Made apply() method to only apply side effects on the model

Check if all expressions in the new_top are bool on a rule_engine level

Adapted tree-sitter to a new Model definition

Now we use .unverse() instead of .transform in
assert_vector_operators_have_partially_evaluated()