rules(minion): add modulo
Add modulo expressions, parsing, and conversion rules to Minion.

The undefinedness semantics for modulo are similar to those for
division: it requires a bubble rule, asserting that for x % y, y!=0, to
make it safe.

Due to their similarity, the test cases for modulo in this patch are
based on the division ones.

See also: #454 (implementation of division rules)

Closes: #456