refactor(rules/minion): only use bool_to_wlit for Not(Lit::Bool)
Rename bool_to_wlit to not_literal_to_wliteral and make it only
applicable for Not(Lit::Bool) expressions.

PR #473 made non negated boolean literals be converted to w-lit in the
solver interface directly, making this rule mostly redundant.

As Not(Lit::Bool) is a nested expression, I still think it is reasonable
to convert this to a watched literal in a rule instead of in the solver
interface. Also, if it were to happen in the solver interface, there is
a chance that the literal could be reified instead, which is not what we
want.

Fixes: #474