fix(tester): remove duplicate solutions when checking against Conjure
Remove duplicate solutions found by Minion before comparing them against
Conjures solutions.

Duplicates are introduced when we remove auxiliary variables from our
solutions. Conjure. For example, `{__0=1,x=1,y=2}` and `{__0=0,x=1,y=2}`
become identical to each other. As Conjure only has this solution once,
the check would fail. This was causing problems in #425.

To fix this, this patch changes the type of solutions from `HashMap` to
`BTreeMap`. This implements `Eq` and `Hash`, allowing us to use
`.unique()` on our list of solutions.

Note that this patch only changes this in the tester; future work should
make all solution fetching code use `BTreeMap`. Once this is done, the
solution checking logic could likely be simplified as it was written the
way it was because `HashMaps` could not be directly compared.