refactor(ast): make `Expr::can_be_undefined` check children
Make `Expr::can_be_undefined` check its children to see if they are
undefined. Rename it to `Expr::is_unsafe()` to better reflect this.

Before this commit, `can_be_undefined` returned true if the expression
could possibly contain an unsafe child. In practice, this made most
recursive expressions possibly undefined.

in (0f3c9a1330a, refactor(rules/minion): apply `sum_eq_to_sumeq` on
`AuxDeclaration`, 2024-11-28), this caused problems with the bubbling of
sums inside quotients.

The expression

```
UnsafeDiv(Sum[1 + 2 + 3],4)
```

was not able to be converted into a `SafeDiv` because it contained a
possibly undefined expression type, `Sum`.

Now, we check inside the sum recursively, and only call it unsafe if it
contains an `UnsafeDiv` or `UnsafeMod`.