refactor: make all Minion expressions flat
Change the type signature of Minion constraints to use `Atom`, `Literal`
and `Name` instead of recursive `Expressions`. This enforces flatness
through typing.

Give these expressions the Flat or Minion prefixes to better reflect the
fact that they are solver-level expressions only.

This commit also contains the following refactors:

  * Simplified and rewrote much of the Minion solver adaptor
    `load_model` code, and moved it to a separate file.

  * Improve documentation of Minion expressions, and add links to their
    Minion documentation

  * Rename some Minion rules to be more consistent, and update their doc
    strings.

  * Remove `use Expr::*` from the Minion rules.

Apart from the rule name changes, these refactors do not change
functionality.

DETAILS

The idea of encoding flatness as types originated in a872e12b0
(rules(minion): add flattening rules for division, 2024-11-16).

This commit adds this strong typing to the rest of the Minion
constraints.

Minion takes flat constraints as input. Encoding this limitation in the
type signature guarantees that our rules obey this, and that nothing is
converted into Minion constraints until it is flat.

Keeping expressions in a solver-generic, higher-level form until they
can be directly converted to a Minion valid constraint is desirable as
they can be rewritten by solver-generic rules (such as normalisation)
for longer.

The prefix used to rename an expression depends on how Minion specific
they are: constraints that could be found in other solvers, such as
SumGeq, are called Flat. Constraints with more Minion specific
semantics, such as DivEqUndefZero, are prefixed with Minion instead.