rules: simplify Gt/Lt to Geq/Leq
* Add normalising rules to simplify Gt to Geq and Lt to Leq.
* Remove redundant Lt / Gt specific Minion rules.
* Add y_plus_k_geq_x_to_ineq rule.

  This is the same as x_keq_y_plus_k_to_ineq but with the sign and
  arguments flipped. This allows gt to convert to ineq instead of
  sumgeq.

Minion has several constraints that deal with Geq and Leq such as
SumLeq and WeightedSumLeq. Before this commit, conversion rules to
SumLeq, SumGeq, and Ineq had special cases to deal with Lt and Gt. By
normalising Gt/Lt to Geq and Leq, Minion rules no longer need to deal
with this special case. This change simplifies some Minion rules, and makes
other completely redundant.

The motivation for this change was to avoid adding a special case for Lt
and Gt in my upcoming conversion rules for Minion's weighted_sumgeq /
weighted_sumleq constraints.