fix(tester): make human rule trace writer blocking
Make human rule trace use a blocking writer.

I hope that this will fix our issues with macos CI in #516.

The failure is that, in some tests, the generated rule trace has some
lines missing from the end of the file compared to the expected rule
trace. Because of this, I suspect that the generated file is being read
before it has fully been written to. If this is the issue, blocking on
write would fix this.