amend! Add absolute values and convert to Minion
Add absolute values and convert to Minion

Add support for absolute values, in particular:

* Parsing from JSON.
* The `Abs` expression variant.
* The `FlatAbsEq` Minion constraint.
* Flattening and conversion to Minion.

As part of this change, a division domain bug is fixed. In our domain
calculation we use truncating division, the default Rust integer
division operator. However, the semantics of Essence Prime specify floor
division.

Truncating division has the effect of rounding towards 0.

Truncating division does `-5 / 2 =-2` and `5/2 = 2`. However, we want
`-5/2 = -3` and `5/2 = 2`.

This caused the domains of variables containing negative divisions to
sometimes be one too high. This caused abs-03 to give different
solutions to Conjure.