cli: add --no-run-solver flag
Add --no-run-solver flag to disable the running of the solver. This
prints the rewritten model as output.

The motivation for this flag is comparing execution time against Savile
Row. In the future we may benchmark more rigorously with the reported
rewrite times output by both. However, this flag would still be useful
for doing ad-hoc performance comparisons during development (e.g. using
  `time` or `hyperfine`).

It could also be useful for debugging the rewriting of larger models.

The model output is not necessarily valid Essence that we can parse
(as it can contain Minion constraints), so might not be too useful
outside of these use-cases.