perf: only log rule not applicable in debug mode
Only run the "rule not applicable" trace when Conjure Oxide is built in
debug mode. For the test file
`optimisations/implies-tautologies-cse/input.eprime`, this reduces
execution time from 5.42s to 1.281s, a 75% improvement.

As this is only printed at the trace debug level, it is unlikely to be
needed in a release build anyways.

DETAILS

Measurements were taken using `hyperfine`, building the project in
release mode. Conjure Oxide was ran using the `--no-run-solver` flag to
mimic default Savile Row behaviour.

Profiling (using `samply`) revealed that 75% of our execution time was
spent in this `trace`, even if we never print it.

For comparison, `savilerow -O0` takes 0.136s +- 0.004. Note that part of
this may be due to us having no implication simplification rules,
meaning that we would need to do more rewrites than Savile Row.

Another expensive function is `Uniplate::contexts`. A rewrite for this
is in progress.