feat(rules): add more `->` tautologies to the partial evaluator
Add the following implication tautologies to the partial evaluator:

  + Totality: `(p->q) \/ (q->p) ~> true`

  + Conditional excluded middle: `(p->q) \/ (p-> !q) ~> true`

These tautologies both involve checking pairs inside an `or`. I
originally thought checking pairs was too expensive, so did not
implement them.

Primarily, Savile Row does not check these, so adding them here is a
good experiment to see how useful they are. Given the frequency of
complex conditionals in modelling, it is likely to give some sort of
benefit.

The implementation in this commit does not change the time complexity to
check disjunctions that do not contain implications - this is still
O(n).

Issue: https://github.com/conjure-cp/conjure-oxide/issues/511