tester: enable `prop_multiple_equally_applicable` rewriter check
Enable the `prop_multiple_equally_applicable` rewriter assertion in the
integration tester. This check panics if a given expression has two
rules of the same priority applicable to it.

All tests pass with this enabled.

For performance reasons, I think it is best to keep this disabled by
default in the CLI.

Issue: #552