refactor(ast)!: refactor `SymbolTable` interface to hide implementation details
Change the `SymbolTable` to be a struct with private members, and create
a new interface to interact with it. This hides the fact that the
`SymbolTable` is a `BTreeMap`.

While the symbol table is a simple `BTreeMap` both before and after this
commit, the implementation of this is likely to change multiple times in
the future. In particular, this commit is preparation for adding
`lettings` and `givens` to the symbol table.