fix(tester): make verbose work again, and run non-verbose tests in parallel
Previous to this commit, VERBOSE=true caused the tests to hang
indefinitely. This commit fixes this, and makes tests without
ACCEPT=true or VERBOSE=true run in parallel again. ACCEPT=true tests
cannot be run in parallel as I get occasional Conjure errors when doing
this (which in turn cause the tests to fail)