feat(ast): add value lettings to symbol table.
Add value lettings (`letting NAME be EXPRESSION`) to the symbol table,
and add methods to mutate and access these.

This commit also moves documentation from the module-level to the
`SymbolTable` struct. Because the `symbol_table` module is private, the
module level documentation wasn't visible in rustdoc. As `SymbolTable`
is re-exported by `ast`, the docs are visible there.

Closes: #522