feat(naive_rewriter): also rewrite expressions inside lettings
Allow rewriting of expressions inside lettings.

The rewriter, on each pass, checks for possible applicable rules in the
lettings then the model. The semantics of rewriting sub-expressions
inside lettings and the model are the same as before.

Rewriting `letting A be EXPENSIVE` (where EXPENSIVE is a complex
expression which needs many rule applications to refine) before we
substitute it should reduce the number of rule applications needed.