refactor: Remove RuleSet order
As discussed with Oz, RuleSet order is only ever used as a tie breaker in very specific circumstances. It's unnecessary and very confusing for the user, so I suggest we remove it.

See this link for the original reasoning for adding it:

https://github.com/conjure-cp/conjure-oxide/wiki/Expression-rewriting%2C-Rules-and-RuleSets#ruleset-ordering

(This doc will need to be updated to avoid confusion)