refactor: simplify naive rewriter using `Biplate<Expr>`
Simplify the naive rewriter logic by implementing `Biplate<Expression>`
for `Model`.

Currently, we have two similar functions to rewrite expressions in value
lettings and the expression tree. By implementing `Biplate<Expression>`
in this commit, the constraints tree and expressions inside the symbol
table can be traversed using the same operation. This makes it possible to
use a single rewriting loop for both value lettings and the constraints
tree.

Due to limitations in Uniplate derive (documented in code comments),
this commit uses manual implementations for the Uniplate / Biplate
traits for `Model` and `SymbolTable`; however, the features needed to
make derive work here have been added to the Uniplate issue tracker, so
we should be able to remove these in the future.