feat: add `Expr::Scope`
Add `Scope` expression variant, allowing submodels to be defined inside a model.

Change the rewriter so that it rewrites expressions inside their
submodels, using that submodels symbol table. This allows rules to apply
to variables defined in a local scope.

This commit is part of on-going work to add lexical scope to Conjure
Oxide, following on from 542c89ee7 (feat!: add `Submodel`, 2025-02-16).