fix(tester): only regenerate expected test outputs when necessary
For test stages that use a model JSON, only regenerate the expected test
outputs if the test stage fails.

BACKGROUND

In 863643260 (feat!: add parent symbol tables (#680), 2025-02-17), we
added an id field to symbol-tables, allowing the reconstruction of
pointers during de-serialisation.

These ids are set by a global variable; therefore, they tend to change
from run to run of the tester. However, they are ignored in the
implementation of `Eq` for Model/SubModel, so this does not affect test
correctness.

After this change, when running the test suite with ACCEPT=true, new
expected output files would be created for all test models containing
different ids.

This caused our git diffs to always contain changes to the ids in every
single test case, cluttering them up.

As id changes do not tests to fail, this commit means that an id change
will not cause a new expected test output file to be generated.