fix(tester): when ACCEPT=true, do not panic on JSON parsing errors
When ACCEPT=true, do not panic when the expected JSON is invalid.

Refactoring the AST data types can cause previously valid expected JSON
to become invalid. Before this commit, this would cause the tester to
panic. Now, instead of panicking, this causes the test to be marked as
failed, causing the expected JSON to be regenerated.