ci: move hygiene checks to `Makefile`, allowing them to be ran locally
Create a Makefile containing our PR hygiene checks, and the commands
necessary to fix it. Refactor the hygiene CI to call this Makefile.

This allows the running of the hygiene tests used in CI locally using
`make check`. Having CI call the Makefile, ensures that the local and CI
versions will always be consistent.

The fix/fix-dirty commands wrap the correct cargo clippy and cargo fmt
flags to auto-fix these issues.

This Makefile is self-documenting, and help can be viewed by running
`make`.