feat: add representation rules for matrices (#728)
Add representations, and using them solve matrices using Minion.

The representation setup added in this commit is a prototype and is
subject to change (due to time constraints). In particular:

  + Nested representations (e.g. for nested sets) or multiple
    representations for a variable are not supported.

  + Error handling is not complete, and some panics could be replaced by
    throwing an error and visa versa.

  + The storage of representations in the symbol table is likely to
    change.

  + Solution translation code involving represented names is hard-coded
    into the Minion adaptor, whereas this should be solver independent.

  + Documentation is incomplete.

INTEGRATION TESTER

As part of this work, it was necessary to modify the integration tester
to support saving abstract literal solutions to JSON, and parsing them
from Conjure. To read solutions emitted by Conjure, we now pass the
Essence solutions files emitted by Conjure through the parser, instead
of directly dealing with the solutions JSON. As each solution is a
seperate file, there are some performance issues with doing this for
models with lots of solutions.

A more ideal solution we would to directly get the list of solutions as
a JSON array in Conjures AST format, and pass that directly to the JSON
parser: however, getting AST JSON output out of `conjure solve` is
currently broken.

FUTURE WORK

Matrix solving will be completed with an upcoming commit adding total
evaluation of the indexing and sluicing of matrix literals.